+++ src/adapters/index.ts
@@ -2,3 +2,4 @@
  * Adapters index
  */
 export { QQAdapter } from "./qq/index";
+export { createAdapter, type PlatformType } from "./factory";